refactor: changing folders name
