+++ internal/adapters/drivers/api/DTOs/create_content_dto.go
